more porting !!
utilize tablist_uuids.json
separate element mixins for later
clean up mixins
