+++ src/main/java/org/polyfrost/vanillahud/mixin/editor/PlayerTabOverlayMixin.java
@@ -1,0 +1,24 @@
+package org.polyfrost.vanillahud.mixin.editor;
+
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import net.minecraft.client.gui.components.PlayerTabOverlay;
+import net.minecraft.client.multiplayer.PlayerInfo;
+import org.polyfrost.oneconfig.api.hud.v1.HudManager;
+import org.polyfrost.vanillahud.util.TabListManager;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+import java.util.List;
+
+@Mixin(PlayerTabOverlay.class)
+public class PlayerTabOverlayMixin {
+    @ModifyReturnValue(method = "getPlayerInfos", at = @At("RETURN"))
+    private List<PlayerInfo> vanillahud$demoPlayers(List<PlayerInfo> original) {
+        if (HudManager.INSTANCE.isEditing()) {
+            TabListManager.INSTANCE.ensureLoaded();
+
+            return TabListManager.INSTANCE.getDevInfo();
+        }
+        return original;
+    }
+}
