+++ src/main/java/org/polyfrost/vanillahud/mixin/elements/GuiMixinHunger.java
@@ -1,0 +1,38 @@
+package org.polyfrost.vanillahud.mixin.elements;
+
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import net.minecraft.client.gui.GuiGraphicsExtractor;
+import net.minecraft.world.entity.player.Player;
+import org.polyfrost.vanillahud.hud.Huds;
+import org.polyfrost.vanillahud.render.HudTransform;
+import org.spongepowered.asm.mixin.Mixin;
+
+//? if >=26.2 {
+import net.minecraft.client.gui.Hud;
+//?} else {
+//import net.minecraft.client.gui.Gui;
+//?}
+
+//? if >=26.2 {
+@Mixin(Hud.class)
+//?} else {
+//@Mixin(Gui.class)
+//?}
+public class GuiMixinHunger {
+    @WrapMethod(
+            //? if < 26 {
+            /*method = "renderFood"
+            *///?} else {
+            method = "extractFood"
+            //?}
+    )
+    private void vanillahud$hunger(
+            GuiGraphicsExtractor graphics, Player player, int yLineBase, int xRight, Operation<Void> original) {
+        if (!Huds.INSTANCE.getHunger().shouldRender()) return;
+
+        HudTransform.begin(graphics, Huds.INSTANCE.getHunger());
+        original.call(graphics, player, yLineBase, xRight);
+        HudTransform.end(graphics);
+    }
+}
